+++ l10n/kn.js
@@ -1,6 +1,7 @@
 OC.L10N.register(
     "fulltextsearch",
     {
-    "Search" : "Search"
+    "Search" : "Search",
+    "Error" : "﻿ತಪ್ಪಾಗಿದೆ"
 },
 "nplurals=2; plural=(n > 1);");