+++ l10n/sq.js
@@ -2,7 +2,9 @@ OC.L10N.register(
     "fulltextsearch",
     {
     "Search" : "Kërko",
-    "Search on %s" : "Kërkoni në %s",
-    "General" : "Të përgjithshme"
+    "General" : "Të përgjithshme",
+    "Error" : "Gabim",
+    "No results" : "Asnjë rezultat",
+    "Search on %s" : "Kërkoni në %s"
 },
 "nplurals=2; plural=(n != 1);");