+++ lib/Command/Index.php
@@ -10,8 +10,6 @@ declare(strict_types=1);
 namespace OCA\FullTextSearch\Command;
 
 use Exception;
-use OC\Core\Command\InterruptedException;
-use OCA\FullTextSearch\ACommandBase;
 use OCA\FullTextSearch\Exceptions\PlatformTemporaryException;
 use OCA\FullTextSearch\Exceptions\TickDoesNotExistException;
 use OCA\FullTextSearch\Model\Index as ModelIndex;
@@ -23,67 +21,73 @@ use OCA\FullTextSearch\Service\PlatformService;
 use OCA\FullTextSearch\Service\ProviderService;
 use OCA\FullTextSearch\Service\RunningService;
 use OCA\FullTextSearch\Tools\Traits\TArrayTools;
+use OCP\Console\Attribute\Argument;
+use OCP\Console\Attribute\AsCommand;
+use OCP\Console\Attribute\Option;
+use OCP\Console\ExitCode;
+use OCP\Console\ISignalHandler;
 use OCP\FullTextSearch\IFullTextSearchProvider;
 use OCP\IUserManager;
 use OutOfBoundsException;
-use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Terminal;
 use Throwable;
 
-class Index extends ACommandBase {
+#[AsCommand(
+	name: 'fulltextsearch:index',
+	description: 'Index files',
+)]
+class Index {
 	use TArrayTools;
 
-	const INDEX_OPTION_NO_READLINE = '_no-readline';
+	public const INDEX_OPTION_NO_READLINE = '_no-readline';
 
-//			'%job:1s%%message:-40s%%current:6s%/%max:6s% [%bar%] %percent:3s%% \n %duration% %infos:-12s% %jvm:-30s%      '
-	const PANEL_RUN = 'run';
-	const PANEL_RUN_LINE_OPTIONS = 'Options: %options%';
-	const PANEL_RUN_LINE_MEMORY = 'Memory: %_memory%';
+	//			'%job:1s%%message:-40s%%current:6s%/%max:6s% [%bar%] %percent:3s%% \n %duration% %infos:-12s% %jvm:-30s%      '
+	public const PANEL_RUN = 'run';
+	public const PANEL_RUN_LINE_OPTIONS = 'Options: %options%';
+	public const PANEL_RUN_LINE_MEMORY = 'Memory: %_memory%';
 
-	const PANEL_INDEX = 'indexing';
-	const PANEL_INDEX_LINE_HEADER = '┌─ Indexing %_paused% ────';
-	const PANEL_INDEX_LINE_ACCOUNT = '│ Provider: <info>%providerName:-20s%</info> Account: <info>%userId%</info>';
-	const PANEL_INDEX_LINE_ACTION = '│ Action: <info>%action%</info>';
-	const PANEL_INDEX_LINE_DOCUMENT = '│ Document: <info>%documentId%</info>';
-	const PANEL_INDEX_LINE_INFO = '│ Info: <info>%info%</info>';
-	const PANEL_INDEX_LINE_TITLE = '│ Title: <info>%title%</info>';
-	const PANEL_INDEX_LINE_CONTENT = '│ Content size: <info>%content%</info>';
-	const PANEL_INDEX_LINE_CHUNK = '│ Chunk: %chunkCurrent:6s%/%chunkTotal%';
-	const PANEL_INDEX_LINE_PROGRESS = '│ Progress: %documentCurrent:6s%/%documentTotal%';
-	const PANEL_INDEX_LINE_FOOTER = '└──';
+	public const PANEL_INDEX = 'indexing';
+	public const PANEL_INDEX_LINE_HEADER = '┌─ Indexing %_paused% ────';
+	public const PANEL_INDEX_LINE_ACCOUNT = '│ Provider: <info>%providerName:-20s%</info> Account: <info>%userId%</info>';
+	public const PANEL_INDEX_LINE_ACTION = '│ Action: <info>%action%</info>';
+	public const PANEL_INDEX_LINE_DOCUMENT = '│ Document: <info>%documentId%</info>';
+	public const PANEL_INDEX_LINE_INFO = '│ Info: <info>%info%</info>';
+	public const PANEL_INDEX_LINE_TITLE = '│ Title: <info>%title%</info>';
+	public const PANEL_INDEX_LINE_CONTENT = '│ Content size: <info>%content%</info>';
+	public const PANEL_INDEX_LINE_CHUNK = '│ Chunk: %chunkCurrent:6s%/%chunkTotal%';
+	public const PANEL_INDEX_LINE_PROGRESS = '│ Progress: %documentCurrent:6s%/%documentTotal%';
+	public const PANEL_INDEX_LINE_FOOTER = '└──';
 
-	const PANEL_RESULT = 'result';
-	const PANEL_RESULT_LINE_HEADER = '┌─ Results ────';
-	const PANEL_RESULT_LINE_RESULT = '│ Result: <info>%resultCurrent:6s%</info>/<info>%resultTotal%</info>';
-	const PANEL_RESULT_LINE_INDEX = '│ Index: <info>%resultIndex%</info>';
-	const PANEL_RESULT_LINE_STATUS = '│ Status: %resultStatusColored%';
-	const PANEL_RESULT_LINE_MESSAGE1 = '│ Message: <info>%resultMessageA%</info>';
-	const PANEL_RESULT_LINE_MESSAGE2 = '│ <info>%resultMessageB%</info>';
-	const PANEL_RESULT_LINE_MESSAGE3 = '│ <info>%resultMessageC%</info>';
-	const PANEL_RESULT_LINE_FOOTER = '└──';
+	public const PANEL_RESULT = 'result';
+	public const PANEL_RESULT_LINE_HEADER = '┌─ Results ────';
+	public const PANEL_RESULT_LINE_RESULT = '│ Result: <info>%resultCurrent:6s%</info>/<info>%resultTotal%</info>';
+	public const PANEL_RESULT_LINE_INDEX = '│ Index: <info>%resultIndex%</info>';
+	public const PANEL_RESULT_LINE_STATUS = '│ Status: %resultStatusColored%';
+	public const PANEL_RESULT_LINE_MESSAGE1 = '│ Message: <info>%resultMessageA%</info>';
+	public const PANEL_RESULT_LINE_MESSAGE2 = '│ <info>%resultMessageB%</info>';
+	public const PANEL_RESULT_LINE_MESSAGE3 = '│ <info>%resultMessageC%</info>';
+	public const PANEL_RESULT_LINE_FOOTER = '└──';
 
-	const PANEL_ERRORS = 'errors';
-	const PANEL_ERRORS_LINE_HEADER = '┌─ Errors ────';
-	const PANEL_ERRORS_LINE_ERRORS = '│ Error: <comment>%errorCurrent:6s%</comment>/<comment>%errorTotal%</comment>';
-	const PANEL_ERRORS_LINE_ERROR_INDEX = '│ Index: <comment>%errorIndex%</comment>';
-	const PANEL_ERRORS_LINE_ERROR_EXCEPTION = '│ Exception: <comment>%errorException%</comment>';
-	const PANEL_ERRORS_LINE_ERROR_MESSAGE1 = '│ Message: <comment>%errorMessageA%</comment>';
-	const PANEL_ERRORS_LINE_ERROR_MESSAGE2 = '│ <comment>%errorMessageB%</comment>';
-	const PANEL_ERRORS_LINE_ERROR_MESSAGE3 = '│ <comment>%errorMessageC%</comment>';
-	const PANEL_ERRORS_LINE_FOOTER = '└──';
+	public const PANEL_ERRORS = 'errors';
+	public const PANEL_ERRORS_LINE_HEADER = '┌─ Errors ────';
+	public const PANEL_ERRORS_LINE_ERRORS = '│ Error: <comment>%errorCurrent:6s%</comment>/<comment>%errorTotal%</comment>';
+	public const PANEL_ERRORS_LINE_ERROR_INDEX = '│ Index: <comment>%errorIndex%</comment>';
+	public const PANEL_ERRORS_LINE_ERROR_EXCEPTION = '│ Exception: <comment>%errorException%</comment>';
+	public const PANEL_ERRORS_LINE_ERROR_MESSAGE1 = '│ Message: <comment>%errorMessageA%</comment>';
+	public const PANEL_ERRORS_LINE_ERROR_MESSAGE2 = '│ <comment>%errorMessageB%</comment>';
+	public const PANEL_ERRORS_LINE_ERROR_MESSAGE3 = '│ <comment>%errorMessageC%</comment>';
+	public const PANEL_ERRORS_LINE_FOOTER = '└──';
 
-	const PANEL_COMMANDS_ROOT = 'root';
-	const PANEL_COMMANDS_ROOT_LINE = '## q:quit ## p:pause ';
-	const PANEL_COMMANDS_PAUSED = 'paused';
-	const PANEL_COMMANDS_PAUSED_LINE = '## q:quit ## u:unpause ## n:next step';
-	const PANEL_COMMANDS_DONE = 'done';
-	const PANEL_COMMANDS_DONE_LINE = '## q:quit';
-	const PANEL_COMMANDS_NAVIGATION = 'navigation';
-	const PANEL_COMMANDS_ERRORS_LINE = '## f:first error ## h/j:prec/next error ## d:delete error ## l:last error';
-	const PANEL_COMMANDS_RESULTS_LINE = '## x:first result ## c/v:prec/next result ## b:last result';
+	public const PANEL_COMMANDS_ROOT = 'root';
+	public const PANEL_COMMANDS_ROOT_LINE = '## q:quit ## p:pause ';
+	public const PANEL_COMMANDS_PAUSED = 'paused';
+	public const PANEL_COMMANDS_PAUSED_LINE = '## q:quit ## u:unpause ## n:next step';
+	public const PANEL_COMMANDS_DONE = 'done';
+	public const PANEL_COMMANDS_DONE_LINE = '## q:quit';
+	public const PANEL_COMMANDS_NAVIGATION = 'navigation';
+	public const PANEL_COMMANDS_ERRORS_LINE = '## f:first error ## h/j:prec/next error ## d:delete error ## l:last error';
+	public const PANEL_COMMANDS_RESULTS_LINE = '## x:first result ## c/v:prec/next result ## b:last result';
 
 
 	/** @var Runner */
@@ -112,33 +116,28 @@ class Index extends ACommandBase {
 		private PlatformService $platformService,
 		private ProviderService $providerService,
 	) {
-		parent::__construct();
-	}
-
-	protected function configure(): void {
-		parent::configure();
-		$this->setName('fulltextsearch:index')
-			 ->setDescription('Index files')
-			 ->addArgument('options', InputArgument::OPTIONAL, 'options')
-			 ->addOption(
-				 'no-readline', 'r', InputOption::VALUE_NONE,
-				 'disable readline - non interactive mode'
-			 );
 	}
 
 
 	/**
 	 * @throws Exception
 	 */
-	protected function execute(InputInterface $input, OutputInterface $output): int {
-		$options = $this->generateIndexOptions($input);
+	public function __invoke(
+		OutputInterface $output,
+		ISignalHandler $signalHandler,
+		#[Argument(name: 'options', description: 'options')]
+		string $optionsJson = '',
+		#[Option(name: 'no-readline', description: 'disable readline - non interactive mode', shortcut: 'r')]
+		bool $noReadline = false,
+	): ExitCode {
+		$options = $this->generateIndexOptions($optionsJson, $noReadline);
 
 		if ($options->getOptionBool(self::INDEX_OPTION_NO_READLINE, false) === false) {
 			/** do not get stuck while waiting interactive input */
 			try {
 				readline_callback_handler_install(
 					'', function () {
-				}
+					}
 				);
 			} catch (Throwable $t) {
 				throw new Exception('Please install php-readline, or use --no-readline');
@@ -149,9 +148,9 @@ class Index extends ACommandBase {
 
 		$this->terminal = new Terminal();
 
-//		$outputStyle = new OutputFormatterStyle('white', 'black', ['bold']);
-//		$output->getFormatter()
-//			   ->setStyle('char', $outputStyle);
+		//		$outputStyle = new OutputFormatterStyle('white', 'black', ['bold']);
+		//		$output->getFormatter()
+		//			   ->setStyle('char', $outputStyle);
 
 		$this->runner = new Runner($this->runningService, 'commandIndex', ['nextStep' => 'n']);
 		$this->runner->onKeyPress([$this, 'onKeyPressed']);
@@ -166,7 +165,7 @@ class Index extends ACommandBase {
 		$this->runner->setInfo('options', json_encode($options));
 
 		try {
-			$this->runner->sourceIsCommandLine($this, $output);
+			$this->runner->sourceIsCommandLine($signalHandler, $output);
 			$this->runner->start();
 
 			if ($options->getOption('errors') === 'reset') {
@@ -200,7 +199,7 @@ class Index extends ACommandBase {
 		$this->runner->setInfo('documentCurrent', 'all');
 		$this->runner->stop();
 
-		return self::SUCCESS;
+		return ExitCode::Success;
 	}
 
 
@@ -323,15 +322,14 @@ class Index extends ACommandBase {
 
 
 	/**
-	 * @param InputInterface $input
+	 * @param string $jsonOptions
+	 * @param bool $noReadline
 	 *
 	 * @return IndexOptions
 	 */
-	private function generateIndexOptions(InputInterface $input): IndexOptions {
-		$jsonOptions = $input->getArgument('options');
-
+	private function generateIndexOptions(string $jsonOptions, bool $noReadline): IndexOptions {
 		$options = [];
-		if (is_string($jsonOptions)) {
+		if ($jsonOptions !== '') {
 			$options = json_decode($jsonOptions, true);
 		}
 
@@ -339,7 +337,7 @@ class Index extends ACommandBase {
 			$options = [];
 		}
 
-		if ($input->getOption('no-readline')) {
+		if ($noReadline) {
 			$options['_no-readline'] = true;
 		}
 
@@ -400,62 +398,62 @@ class Index extends ACommandBase {
 
 		$this->cliService->createPanel(
 			self::PANEL_INDEX, [
-								 self::PANEL_INDEX_LINE_HEADER,
-								 self::PANEL_INDEX_LINE_ACTION,
-								 self::PANEL_INDEX_LINE_ACCOUNT,
-								 self::PANEL_INDEX_LINE_DOCUMENT,
-								 self::PANEL_INDEX_LINE_INFO,
-								 self::PANEL_INDEX_LINE_TITLE,
-								 self::PANEL_INDEX_LINE_CONTENT,
-								 self::PANEL_INDEX_LINE_CHUNK,
-								 self::PANEL_INDEX_LINE_PROGRESS,
-								 self::PANEL_INDEX_LINE_FOOTER,
-							 ]
+				self::PANEL_INDEX_LINE_HEADER,
+				self::PANEL_INDEX_LINE_ACTION,
+				self::PANEL_INDEX_LINE_ACCOUNT,
+				self::PANEL_INDEX_LINE_DOCUMENT,
+				self::PANEL_INDEX_LINE_INFO,
+				self::PANEL_INDEX_LINE_TITLE,
+				self::PANEL_INDEX_LINE_CONTENT,
+				self::PANEL_INDEX_LINE_CHUNK,
+				self::PANEL_INDEX_LINE_PROGRESS,
+				self::PANEL_INDEX_LINE_FOOTER,
+			]
 		);
 
 		$this->cliService->createPanel(
 			self::PANEL_RESULT, [
-								  self::PANEL_RESULT_LINE_HEADER,
-								  self::PANEL_RESULT_LINE_RESULT,
-								  self::PANEL_RESULT_LINE_INDEX,
-								  self::PANEL_RESULT_LINE_STATUS,
-								  self::PANEL_RESULT_LINE_MESSAGE1,
-								  self::PANEL_RESULT_LINE_MESSAGE2,
-								  self::PANEL_RESULT_LINE_MESSAGE3,
-								  self::PANEL_RESULT_LINE_FOOTER,
-							  ]
+				self::PANEL_RESULT_LINE_HEADER,
+				self::PANEL_RESULT_LINE_RESULT,
+				self::PANEL_RESULT_LINE_INDEX,
+				self::PANEL_RESULT_LINE_STATUS,
+				self::PANEL_RESULT_LINE_MESSAGE1,
+				self::PANEL_RESULT_LINE_MESSAGE2,
+				self::PANEL_RESULT_LINE_MESSAGE3,
+				self::PANEL_RESULT_LINE_FOOTER,
+			]
 		);
 
 		$this->cliService->createPanel(
 			self::PANEL_ERRORS, [
-								  self::PANEL_ERRORS_LINE_HEADER,
-								  self::PANEL_ERRORS_LINE_ERRORS,
-								  self::PANEL_ERRORS_LINE_ERROR_INDEX,
-								  self::PANEL_ERRORS_LINE_ERROR_EXCEPTION,
-								  self::PANEL_ERRORS_LINE_ERROR_MESSAGE1,
-								  self::PANEL_ERRORS_LINE_ERROR_MESSAGE2,
-								  self::PANEL_ERRORS_LINE_ERROR_MESSAGE3,
-								  self::PANEL_ERRORS_LINE_FOOTER,
-							  ]
+				self::PANEL_ERRORS_LINE_HEADER,
+				self::PANEL_ERRORS_LINE_ERRORS,
+				self::PANEL_ERRORS_LINE_ERROR_INDEX,
+				self::PANEL_ERRORS_LINE_ERROR_EXCEPTION,
+				self::PANEL_ERRORS_LINE_ERROR_MESSAGE1,
+				self::PANEL_ERRORS_LINE_ERROR_MESSAGE2,
+				self::PANEL_ERRORS_LINE_ERROR_MESSAGE3,
+				self::PANEL_ERRORS_LINE_FOOTER,
+			]
 		);
 
 		$this->cliService->createPanel(
 			self::PANEL_COMMANDS_PAUSED, [
-										   self::PANEL_COMMANDS_PAUSED_LINE
-									   ]
+				self::PANEL_COMMANDS_PAUSED_LINE
+			]
 		);
 
 		$this->cliService->createPanel(
 			self::PANEL_COMMANDS_ROOT, [
-										 self::PANEL_COMMANDS_ROOT_LINE
-									 ]
+				self::PANEL_COMMANDS_ROOT_LINE
+			]
 		);
 
 		$this->cliService->createPanel(
 			self::PANEL_COMMANDS_NAVIGATION, [
-											   self::PANEL_COMMANDS_RESULTS_LINE,
-											   self::PANEL_COMMANDS_ERRORS_LINE
-										   ]
+				self::PANEL_COMMANDS_RESULTS_LINE,
+				self::PANEL_COMMANDS_ERRORS_LINE
+			]
 		);
 
 		$this->cliService->initDisplay();
@@ -767,18 +765,4 @@ class Index extends ACommandBase {
 
 		$this->displayError();
 	}
-
-
-	/**
-	 * @throws TickDoesNotExistException
-	 */
-	public function abort(): void {
-		try {
-			$this->abortIfInterrupted();
-		} catch (InterruptedException $e) {
-			$this->runner->stop();
-			exit();
-		}
-	}
 }
-