+++ lib/Command/Test.php
@@ -10,9 +10,7 @@ declare(strict_types=1);
 namespace OCA\FullTextSearch\Command;
 
 use Exception;
-use OC\Core\Command\InterruptedException;
 use OC\FullTextSearch\Model\DocumentAccess;
-use OCA\FullTextSearch\ACommandBase;
 use OCA\FullTextSearch\Exceptions\InterruptException;
 use OCA\FullTextSearch\Exceptions\ProviderDoesNotExistException;
 use OCA\FullTextSearch\Exceptions\ProviderIsNotCompatibleException;
@@ -29,17 +27,21 @@ use OCA\FullTextSearch\Service\PlatformService;
 use OCA\FullTextSearch\Service\ProviderService;
 use OCA\FullTextSearch\Service\RunningService;
 use OCA\FullTextSearch\Service\TestService;
-use OCP\AppFramework\QueryException;
+use OCP\Console\Attribute\AsCommand;
+use OCP\Console\Attribute\Option;
+use OCP\Console\ExitCode;
+use OCP\Console\IOutput;
+use OCP\Console\ISignalHandler;
 use OCP\FullTextSearch\IFullTextSearchPlatform;
 use OCP\FullTextSearch\IFullTextSearchProvider;
 use OCP\FullTextSearch\Model\IDocumentAccess;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
+use Psr\Container\ContainerExceptionInterface;
 
-class Test extends ACommandBase {
-
-	const DELAY_STABILIZE_PLATFORM = 3;
+#[AsCommand(
+	name: 'fulltextsearch:test',
+	description: 'Testing the platform setup',
+)]
+class Test {
 
 	private Runner $runner;
 
@@ -50,36 +52,30 @@ class Test extends ACommandBase {
 		private IndexService $indexService,
 		private TestService $testService,
 	) {
-		parent::__construct();
-	}
-
-	protected function configure(): void {
-		parent::configure();
-		$this->setName('fulltextsearch:test')
-			 ->setDescription('Testing the platform setup')
-			 ->addOption('json', 'j', InputOption::VALUE_NONE, 'return result as JSON')
-			 ->addOption(
-				 'platform_delay', 'd', InputOption::VALUE_REQUIRED,
-				 'change DELAY_STABILIZE_PLATFORM'
-			 );
 	}
 
 
 	/**
 	 * @throws Exception
 	 */
-	protected function execute(InputInterface $input, OutputInterface $output): int {
-		$platformDelay = ($input->getOption('platform_delay') > 0) ? (int)$input->getOption(
-			'platform_delay'
-		) : self::DELAY_STABILIZE_PLATFORM;
-
+	public function __invoke(
+		IOutput $output,
+		ISignalHandler $signalHandler,
+		#[Option(description: 'delay between test steps in seconds')]
+		string $platformDelay = '3',
+	): ExitCode {
+		$delay = filter_var($platformDelay, FILTER_VALIDATE_INT, ['options' => ['min_range' => 0]]);
+		if ($delay === false) {
+			$output->writeln('platformDelay must be a nonnegative integer');
+			return ExitCode::Invalid;
+		}
 		$this->output($output, '.Testing your current setup:');
 
 		try {
 			$testProvider = $this->testCreatingProvider($output);
 			$this->testMockedProvider($output, $testProvider);
 			$testPlatform = $this->testLoadingPlatform($output);
-			$this->testLockingProcess($output, $testPlatform, $testProvider);
+			$this->testLockingProcess($output, $signalHandler, $testPlatform, $testProvider);
 		} catch (Exception $e) {
 			$this->outputResult($output, false);
 			throw $e;
@@ -87,14 +83,14 @@ class Test extends ACommandBase {
 
 		try {
 			$this->testResetTest($output, $testProvider);
-			$this->pause($output, $platformDelay);
+			$this->pause($output, $delay);
 			$this->testInitIndexing($output, $testPlatform);
 			$this->testIndexingDocuments($output, $testPlatform, $testProvider);
-			$this->pause($output, $platformDelay);
+			$this->pause($output, $delay);
 			$this->testContentLicense($output, $testPlatform);
 			$this->testSearchSimple($output, $testPlatform, $testProvider);
 			$this->testUpdatingDocumentsAccess($output, $testPlatform, $testProvider);
-			$this->pause($output, $platformDelay);
+			$this->pause($output, $delay);
 			$this->testSearchAccess($output, $testPlatform, $testProvider);
 			$this->testSearchShare($output, $testPlatform, $testProvider);
 
@@ -111,14 +107,14 @@ class Test extends ACommandBase {
 
 		$this->output($output, '', true);
 
-		return 0;
+		return ExitCode::Success;
 	}
 
 
 	/**
 	 * @return IFullTextSearchProvider
 	 * @throws ProviderIsNotCompatibleException
-	 * @throws QueryException
+	 * @throws ContainerExceptionInterface
 	 * @throws ProviderDoesNotExistException
 	 * @throws ProviderIsNotUniqueException
 	 */
@@ -133,11 +129,11 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param string $line
 	 * @param bool $isNewLine
 	 */
-	private function output(OutputInterface $output, string $line, bool $isNewLine = true) {
+	private function output(IOutput $output, string $line, bool $isNewLine = true) {
 		if ($isNewLine) {
 			$output->write(' ', true);
 		}
@@ -147,10 +143,10 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param bool $result
 	 */
-	private function outputResult(OutputInterface $output, bool $result) {
+	private function outputResult(IOutput $output, bool $result) {
 		$isNewLine = false;
 		$line = $this->convertBoolToLine($result, $isNewLine);
 
@@ -175,15 +171,15 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 *
 	 * @return IFullTextSearchProvider
 	 * @throws ProviderDoesNotExistException
 	 * @throws ProviderIsNotCompatibleException
 	 * @throws ProviderIsNotUniqueException
-	 * @throws QueryException
+	 * @throws ContainerExceptionInterface
 	 */
-	private function testCreatingProvider(OutputInterface $output): IFullTextSearchProvider {
+	private function testCreatingProvider(IOutput $output): IFullTextSearchProvider {
 		$this->output($output, 'Creating mocked content provider.');
 		$testProvider = $this->generateMockProvider();
 		$this->outputResult($output, true);
@@ -193,28 +189,28 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchProvider $testProvider
 	 */
 	private function testMockedProvider(
-		OutputInterface $output, IFullTextSearchProvider $testProvider
+		IOutput $output, IFullTextSearchProvider $testProvider,
 	) {
 		$this->output($output, 'Testing mocked provider: get indexable documents.');
 		$testProvider->setIndexOptions(new IndexOptions());
-		$indexableDocuments =
-			$testProvider->generateIndexableDocuments(TestService::DOCUMENT_USER1, '');
+		$indexableDocuments
+			= $testProvider->generateIndexableDocuments(TestService::DOCUMENT_USER1, '');
 		$this->output($output, '(' . sizeof($indexableDocuments) . ' items)', false);
 		$this->outputResult($output, true);
 	}
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 *
 	 * @return IFullTextSearchPlatform
 	 * @throws Exception
 	 */
-	private function testLoadingPlatform(OutputInterface $output): IFullTextSearchPlatform {
+	private function testLoadingPlatform(IOutput $output): IFullTextSearchPlatform {
 		$this->output($output, 'Loading search platform.');
 		$wrapper = $this->platformService->getPlatform();
 		$testPlatform = $wrapper->getPlatform();
@@ -224,7 +220,7 @@ class Test extends ACommandBase {
 
 		$this->output($output, 'Testing search platform.');
 		if (!$testPlatform->testPlatform()) {
-			throw new Exception ('Search platform (' . $testPlatform->getName() . ') down ?');
+			throw new Exception('Search platform (' . $testPlatform->getName() . ') down ?');
 		}
 		$this->outputResult($output, true);
 
@@ -233,19 +229,20 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
+	 * @param ISignalHandler $signalHandler
 	 * @param IFullTextSearchPlatform $testPlatform
 	 * @param IFullTextSearchProvider $testProvider
 	 *
 	 * @throws RunnerAlreadyUpException
 	 */
 	private function testLockingProcess(
-		OutputInterface $output, IFullTextSearchPlatform $testPlatform,
-		IFullTextSearchProvider $testProvider
+		IOutput $output, ISignalHandler $signalHandler, IFullTextSearchPlatform $testPlatform,
+		IFullTextSearchProvider $testProvider,
 	) {
 		$this->output($output, 'Locking process');
 		$this->runner = new Runner($this->runningService, 'test');
-		$this->runner->sourceIsCommandLine($this, $output);
+		$this->runner->sourceIsCommandLine($signalHandler, $output);
 		$this->runner->start();
 		$this->indexService->setRunner($this->runner);
 		$testPlatform->setRunner($this->runner);
@@ -255,12 +252,12 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchProvider $testProvider
 	 *
 	 * @throws Exception
 	 */
-	private function testResetTest(OutputInterface $output, IFullTextSearchProvider $testProvider
+	private function testResetTest(IOutput $output, IFullTextSearchProvider $testProvider,
 	) {
 		$this->output($output, 'Removing test.');
 		$this->indexService->resetIndex($testProvider->getId());
@@ -269,10 +266,10 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchPlatform $testPlatform
 	 */
-	private function testInitIndexing(OutputInterface $output, IFullTextSearchPlatform $testPlatform
+	private function testInitIndexing(IOutput $output, IFullTextSearchPlatform $testPlatform,
 	) {
 		$this->output($output, 'Initializing index mapping.');
 		$testPlatform->initializeIndex();
@@ -281,15 +278,15 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchPlatform $testPlatform
 	 * @param IFullTextSearchProvider $testProvider
 	 *
 	 * @throws Exception
 	 */
 	private function testIndexingDocuments(
-		OutputInterface $output, IFullTextSearchPlatform $testPlatform,
-		IFullTextSearchProvider $testProvider
+		IOutput $output, IFullTextSearchPlatform $testPlatform,
+		IFullTextSearchProvider $testProvider,
 	) {
 		$this->output($output, 'Indexing generated documents.');
 		$options = new IndexOptions(
@@ -305,13 +302,13 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchPlatform $testPlatform
 	 *
 	 * @throws Exception
 	 */
 	private function testContentLicense(
-		OutputInterface $output, IFullTextSearchPlatform $testPlatform
+		IOutput $output, IFullTextSearchPlatform $testPlatform,
 	) {
 
 		try {
@@ -341,15 +338,15 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchPlatform $testPlatform
 	 * @param IFullTextSearchProvider $testProvider
 	 *
 	 * @throws Exception
 	 */
 	private function testSearchSimple(
-		OutputInterface $output, IFullTextSearchPlatform $testPlatform,
-		IFullTextSearchProvider $testProvider
+		IOutput $output, IFullTextSearchPlatform $testPlatform,
+		IFullTextSearchProvider $testProvider,
 	) {
 
 		$this->output($output, 'Searching basic keywords:');
@@ -363,7 +360,7 @@ class Test extends ACommandBase {
 		);
 		$this->search(
 			$output, $testPlatform, $testProvider, $access, 'document is a simple test',
-//			[TestService::DOCUMENT_TYPE_SIMPLE]
+			//			[TestService::DOCUMENT_TYPE_SIMPLE]
 			[TestService::DOCUMENT_TYPE_SIMPLE, TestService::DOCUMENT_TYPE_LICENSE]
 		);
 		$this->search(
@@ -410,20 +407,20 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchPlatform $testPlatform
 	 * @param IFullTextSearchProvider $testProvider
 	 *
 	 * @throws Exception
 	 */
 	private function testUpdatingDocumentsAccess(
-		OutputInterface $output, IFullTextSearchPlatform $testPlatform,
-		IFullTextSearchProvider $testProvider
+		IOutput $output, IFullTextSearchPlatform $testPlatform,
+		IFullTextSearchProvider $testProvider,
 	) {
 		$this->output($output, 'Updating documents access.');
 		$options = new IndexOptions(
 			[
-				'provider'                            => TestProvider::TEST_PROVIDER_ID,
+				'provider' => TestProvider::TEST_PROVIDER_ID,
 				TestService::DOCUMENT_INDEXING_OPTION => TestService::DOCUMENT_INDEXING_ACCESS
 			]
 		);
@@ -436,15 +433,15 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchPlatform $platform
 	 * @param IFullTextSearchProvider $provider
 	 *
 	 * @throws Exception
 	 */
 	private function testSearchAccess(
-		OutputInterface $output, IFullTextSearchPlatform $platform,
-		IFullTextSearchProvider $provider
+		IOutput $output, IFullTextSearchPlatform $platform,
+		IFullTextSearchProvider $provider,
 	) {
 		$this->output($output, 'Searching with group access rights:');
 
@@ -468,15 +465,15 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchPlatform $platform
 	 * @param IFullTextSearchProvider $provider
 	 *
 	 * @throws Exception
 	 */
 	private function testSearchShare(
-		OutputInterface $output, IFullTextSearchPlatform $platform,
-		IFullTextSearchProvider $provider
+		IOutput $output, IFullTextSearchPlatform $platform,
+		IFullTextSearchProvider $provider,
 	) {
 
 		$this->output($output, 'Searching with share rights:');
@@ -489,11 +486,11 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 *
 	 * @throws TickDoesNotExistException
 	 */
-	private function testUnlockingProcess(OutputInterface $output) {
+	private function testUnlockingProcess(IOutput $output) {
 		$this->output($output, 'Unlocking process');
 		$this->runner->stop();
 		$this->outputResult($output, true);
@@ -501,7 +498,7 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchPlatform $testPlatform
 	 * @param IFullTextSearchProvider $testProvider
 	 * @param IDocumentAccess $access
@@ -512,9 +509,9 @@ class Test extends ACommandBase {
 	 * @throws Exception
 	 */
 	private function search(
-		OutputInterface $output, IFullTextSearchPlatform $testPlatform,
+		IOutput $output, IFullTextSearchPlatform $testPlatform,
 		IFullTextSearchProvider $testProvider,
-		IDocumentAccess $access, string $search, array $expected, string $moreOutput = ''
+		IDocumentAccess $access, string $search, array $expected, string $moreOutput = '',
 	) {
 		$this->output(
 			$output,
@@ -541,7 +538,7 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchPlatform $testPlatform
 	 * @param IFullTextSearchProvider $testProvider
 	 * @param array $groups
@@ -550,8 +547,8 @@ class Test extends ACommandBase {
 	 * @throws Exception
 	 */
 	private function searchGroups(
-		OutputInterface $output, IFullTextSearchPlatform $testPlatform,
-		IFullTextSearchProvider $testProvider, array $groups, array $expected
+		IOutput $output, IFullTextSearchPlatform $testPlatform,
+		IFullTextSearchProvider $testProvider, array $groups, array $expected,
 	) {
 
 		$access = new DocumentAccess();
@@ -566,7 +563,7 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param IFullTextSearchPlatform $testPlatform
 	 * @param IFullTextSearchProvider $testProvider
 	 * @param string $user
@@ -575,8 +572,8 @@ class Test extends ACommandBase {
 	 * @throws Exception
 	 */
 	private function searchUsers(
-		OutputInterface $output, IFullTextSearchPlatform $testPlatform,
-		IFullTextSearchProvider $testProvider, string $user, array $expected
+		IOutput $output, IFullTextSearchPlatform $testPlatform,
+		IFullTextSearchProvider $testProvider, string $user, array $expected,
 	) {
 		$access = new DocumentAccess();
 		$access->setViewerId($user);
@@ -608,12 +605,12 @@ class Test extends ACommandBase {
 
 
 	/**
-	 * @param OutputInterface $output
+	 * @param IOutput $output
 	 * @param int $s
 	 *
 	 * @throws InterruptException
 	 */
-	private function pause(OutputInterface $output, int $s) {
+	private function pause(IOutput $output, int $s) {
 		$this->output($output, 'Pausing ' . $s . ' seconds');
 
 		for ($i = 1; $i <= $s; $i++) {
@@ -627,18 +624,4 @@ class Test extends ACommandBase {
 		$this->outputResult($output, true);
 	}
 
-
-	/**
-	 * @throws TickDoesNotExistException
-	 */
-	public function abort() {
-		try {
-			$this->abortIfInterrupted();
-		} catch (InterruptedException $e) {
-			$this->runner->stop();
-			exit();
-		}
-	}
-
 }
-