+++ lib/Service/ConfigService.php
@@ -33,10 +33,10 @@ class ConfigService {
 	}
 
 	public function setConfig(array $save): void {
-		foreach(array_keys($save) as $k) {
-			switch($k) {
+		foreach (array_keys($save) as $k) {
+			switch ($k) {
 				case ConfigLexicon::APP_NAVIGATION:
-					$this->appConfig->setAppValueBool($k, $save[$k]);
+					$this->appConfig->setAppValueBool($k, (bool)$save[$k]);
 					break;
 
 				case ConfigLexicon::SEARCH_PLATFORM: